Fix clearAll and isolate add function
