+++ src/clearAll.js
@@ -1,6 +1,6 @@
-export default function clearAll(element) {
-  if (JSON.parse(localStorage.getItem('list'))) {let list = JSON.parse(localStorage.getItem('list'));
-  element.addEventListener('click', () => {
+export default function clearAll() {
+  if (JSON.parse(localStorage.getItem('list'))) {
+    let list = JSON.parse(localStorage.getItem('list'));
     /* eslint-disable */
     const doubleCheck = confirm('Are you sure?');
     /* eslint-enable */
@@ -13,5 +13,5 @@ export default function clearAll(element) {
         /* eslint-enable */
       }
     }
-  });}
+  }
 }
